Moves portfolio to its own page.
